backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0.0
openai>=1.0.0
pillow>=10.0.0
pandas>=2.0.0
catboost>=1.2.0
scikit-learn>=1.3.0
pandas>=2.0
